requirements.txt
backcall==0.1.0
decorator==4.4.2
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
mysql-connector==2.2.9
numpy==1.18.4
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.3
